Move SolverInstallPlan to its own module.
This is a preparatory commit for giving SolverInstallPlan
its own type.  We first start by moving the type synonyms
for SolverInstallPlan into their own module, and update
module references to point to them.

TODO: Maybe this module should go in the Solver hierarchy
rather than the client hierarchy?
