+++ cabal-install/Distribution/Client/InstallPlan.hs
@@ -15,17 +15,14 @@
 -----------------------------------------------------------------------------
 module Distribution.Client.InstallPlan (
   InstallPlan,
-  SolverInstallPlan,
   GenericInstallPlan,
   PlanPackage,
-  SolverPlanPackage,
   GenericPlanPackage(..),
 
   -- * Operations on 'InstallPlan's
   new,
   toList,
   mapPreservingGraph,
-  configureInstallPlan,
 
   ready,
   processing,
@@ -63,16 +60,13 @@ import Distribution.Package
          , HasUnitId(..), UnitId(..) )
 import Distribution.Client.Types
          ( BuildSuccess, BuildFailure
-         , ConfiguredPackage(..), ConfiguredId(..)
+         , ConfiguredPackage(..)
          , UnresolvedPkgLoc
          , GenericReadyPackage(..) )
 import Distribution.Version
          ( Version )
 import Distribution.Simple.PackageIndex
          ( PackageIndex )
-import qualified Distribution.Simple.Configure as Configure
-import qualified Distribution.Simple.Setup as Cabal
-import qualified Distribution.PackageDescription as PD
 import qualified Distribution.Simple.PackageIndex as PackageIndex
 import qualified Distribution.Client.PlanIndex as PlanIndex
 import Distribution.Text
@@ -82,7 +76,6 @@ import           Distribution.Solver.Types.ComponentDeps (ComponentDeps)
 import qualified Distribution.Solver.Types.ComponentDeps as CD
 import           Distribution.Solver.Types.PackageFixedDeps
 import           Distribution.Solver.Types.Settings
-import           Distribution.Solver.Types.SolverPackage
 
 -- TODO: Need this when we compute final UnitIds
 -- import qualified Distribution.Simple.Configure as Configure
@@ -170,10 +163,6 @@ type PlanPackage = GenericPlanPackage
                    InstalledPackageInfo (ConfiguredPackage UnresolvedPkgLoc)
                    BuildSuccess BuildFailure
 
-type SolverPlanPackage = GenericPlanPackage
-                         InstalledPackageInfo (SolverPackage UnresolvedPkgLoc)
-                         BuildSuccess BuildFailure
-
 instance (Package ipkg, Package srcpkg) =>
          Package (GenericPlanPackage ipkg srcpkg iresult ifailure) where
   packageId (PreExisting ipkg)     = packageId ipkg
@@ -228,57 +217,6 @@ planPkgOf plan v =
                                    (planPkgIdOf plan v) of
       Just pkg -> pkg
       Nothing  -> error "InstallPlan: internal error: planPkgOf lookup failed"
-
--- | 'GenericInstallPlan' that the solver produces.  We'll "run this" in
--- order to compute the 'UnitId's for everything we want to build.
-type SolverInstallPlan = GenericInstallPlan
-                         InstalledPackageInfo (SolverPackage UnresolvedPkgLoc)
-                         -- Technically, these are not used here, but
-                         -- setting the type this way makes it easier
-                         -- to run some operations.
-                         BuildSuccess BuildFailure
-
--- | Conversion of 'SolverInstallPlan' to 'InstallPlan'.
--- Similar to 'elaboratedInstallPlan'
-configureInstallPlan :: SolverInstallPlan -> InstallPlan
-configureInstallPlan solverPlan =
-    flip mapPreservingGraph solverPlan $ \mapDep planpkg ->
-      case planpkg of
-        PreExisting pkg ->
-          PreExisting pkg
-
-        Configured  pkg ->
-          Configured (configureSolverPackage mapDep pkg)
-
-        _ -> error "configureInstallPlan: unexpected package state"
-  where
-    configureSolverPackage :: (UnitId -> UnitId)
-                           -> SolverPackage UnresolvedPkgLoc
-                           -> ConfiguredPackage UnresolvedPkgLoc
-    configureSolverPackage mapDep spkg =
-      ConfiguredPackage {
-        confPkgId = SimpleUnitId
-                  $ Configure.computeComponentId
-                        Cabal.NoFlag
-                        (packageId spkg)
-                        (PD.CLibName (display (pkgName (packageId spkg))))
-                        -- TODO: this is a hack that won't work for Backpack.
-                        (map ((\(SimpleUnitId cid0) -> cid0) . confInstId)
-                             (CD.libraryDeps deps))
-                        (solverPkgFlags spkg),
-        confPkgSource = solverPkgSource spkg,
-        confPkgFlags  = solverPkgFlags spkg,
-        confPkgStanzas = solverPkgStanzas spkg,
-        confPkgDeps   = deps
-      }
-      where
-        deps = fmap (map (configureSolverId mapDep)) (solverPkgDeps spkg)
-
-    configureSolverId mapDep sid =
-      ConfiguredId {
-        confSrcId  = packageId sid, -- accurate!
-        confInstId = mapDep (installedUnitId sid)
-      }
 
 -- | 'GenericInstallPlan' specialised to most commonly used types.
 type InstallPlan = GenericInstallPlan
